Move stories to be next to their components.
