+++ .storybook/config.js
@@ -1,8 +1,8 @@
 import { configure } from "@kadira/storybook"
 
 function loadStories() {
-  require("../stories/artwork")
-  require("../stories/artwork_grid")
+  require("../src/components/__stories__/artwork")
+  require("../src/components/__stories__/artwork_grid")
 }
 
 configure(loadStories, module);
