+++ src/components/__stories__/artwork_grid.tsx
@@ -2,10 +2,10 @@ import * as React from "react"
 import * as Relay from "react-relay"
 import { storiesOf, action } from "@kadira/storybook"
 
-import ArtworkGrid from "../src/components/artwork_grid"
+import ArtworkGrid from "../artwork_grid"
 
-import { artsyNetworkLayer } from "../src/relay/config"
-import { ArtistQueryConfig } from "../src/relay/root_queries"
+import { artsyNetworkLayer } from "../../relay/config"
+import { ArtistQueryConfig } from "../../relay/root_queries"
 
 function GridExample(props: { artistID: string }) {
   Relay.injectNetworkLayer(artsyNetworkLayer())
